# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo List Api",
        "description": "Demo Todo List Api",
        "contact": {
            "email": "joshuaamarfio1@gmail.com"
        },
        "version": "1.0.2"
    },
    "paths": {
...
        "/todo-api/public/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "operationId": "1700c4319ed0a7746c8a54637ea15fe0",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Joshua A"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "joshua@email.com"
                                                    },
                                                    "password": {
                                                        "type": "string",
                                                        "example": "hasedValue"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2022-12-20T10:50:13.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2022-12-20T10:50:13.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add new user",
                "operationId": "5d238ea07fb2740accd4cdc11bd3c756",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "",
                                    "email": "",
                                    "password": ""
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "user added successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Joshua A"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "joshua@email.com"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "417": {
                        "description": "not as expected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "number",
                                            "example": "417"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "error message"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "stack trace string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}